feat: Refactored RulesEngine to improve mutations performance
