+++ src/rules-engine/types.ts
@@ -1,3 +1,5 @@
+import { Entity } from '../types'
+
 export type ResourceData = {
   data: { [k: string]: any }
   resource: { id: string; [k: string]: any }
@@ -83,4 +85,11 @@ export interface Engine {
    * @returns An array of RuleFinding
    */
   processRule: (rule: Rule, data: any) => Promise<RuleFinding[]>
+
+  /**
+   * Transforms RuleFinding array into a mutation array for GraphQL
+   * @param findings resulted findings during rules execution
+   * @returns {Entity[]} Array of generated mutations
+   */
+  prepareMutations: (findings: RuleFinding[]) => Entity[]
 }
